+++ src/data/task.ts
@@ -1,3 +1,4 @@
+import type moment from "moment";
 import type { DueDate } from "../api/domain/dueDate";
 import type { Project } from "../api/domain/project";
 import type { Section } from "../api/domain/section";
@@ -5,6 +6,7 @@ import type { Priority, TaskId } from "../api/domain/task";
 
 export type Task = {
   id: TaskId,
+  createdAt: string,
 
   content: string,
   description: string,